+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Tap letter hint word again to hide word.
 - Move word to other collection.
 - Fix bug where editing resets progress.
+- Tap word hints area to show next word.
+- Hide hint box when showing answer.
 
 ## 1.3.0 - July 26, 2023